Allow debcomprt to define env var configs vs .conf
A make variable called PREPROCESS_ALIASES can be passed in with any
value. This directs make to not use the project configuration file but
rather it assumes the needed configurations will be in the env.

SUDOER_PASSWORD_CRYPT need to be changed to _SUDOER_PASSWORD_CRYPT. Any
configuration expected from the env will need to have its respective
makefile variable appended with the _ prefix.